feat: integrate remove category
