+++ src/services/categories/index.ts
@@ -6,23 +6,22 @@ export async function fetchCategories() {
     { headers: { Authorization: `Bearer ${sessionStorage.getItem('jwt')}` } },
   );
 
-  return result.data.map((category: { id: string; name: string }) => ({
-    id: category.id,
-    name: category.name,
-  }));
+  return result.data;
 }
 
-export async function createCategory(body: { name: string }) {
-  const result = await axios.post(
+export async function createCategory(name: string) {
+  await axios.post(
     `${import.meta.env.VITE_BACKEND_ROUTE}/create-category`,
-    body,
+    { name },
     {
       headers: { Authorization: `Bearer ${sessionStorage.getItem('jwt')}` },
     },
   );
+}
 
-  return result.data.map((category: { id: string; name: string }) => ({
-    id: category.id,
-    name: category.name,
-  }));
+export async function removeCategory(id: string) {
+  await axios.delete(`${import.meta.env.VITE_BACKEND_ROUTE}/delete-category`, {
+    headers: { Authorization: `Bearer ${sessionStorage.getItem('jwt')}` },
+    data: { id },
+  });
 }
